cleaning

+++ app/brutforce.cpp
@@ -86,7 +86,7 @@
 // };
 
 
-int main(size_t argc, char* argv[])
+int main(int argc, char* argv[])
 {
     // eoParser argparser(argc, argv);
